requirements.txt
TelegramBotAPI
requests 
shelve
mysql-connector-python
datetime
pytz
timezonefinder
